third_party/WebKit/LayoutTests/fast/dom/shadow/inner-scope-important-wins-expected.txt

CONSOLE WARNING: line 87: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Outer scope !important rules wins with higher specificity (v0 bug compatibility).

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getComputedStyle(target).color is "rgb(255, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
